+++ src/Console/Commands/SyncTemplatesCommand.php
@@ -229,7 +229,7 @@ class SyncTemplatesCommand extends Command
     public function handle()
     {
         $this->basePath = base_path();
-        $this->stubPath = base_path('mic-sole-laravel-gen/src/Templates/');
+        $this->stubPath = \MicSoleLaravelGen\Providers\MicSoleLaravelGenServiceProvider::getTemplatesPath() . '/';
 
         $this->info('🔄 Syncing templates from generated files...');
         $this->newLine();